Growth

Default

\[PC_j = PC_j^{max} \cdot \gamma^{nut}_j \cdot \gamma^{light}_j \cdot f^{temp}_j \cdot \gamma^{CO2}_j\]

Agate.default_PCⱼ — Method

PCⱼ = PCⱼᵐᵃˣ * γⱼⁿᵘᵗ *  γⱼˡⁱᵍʰᵗ * fⱼᵗᵉᵐᵖ *  γⱼᶜᵒ²

Carbon-specific growth rate for plankton j (Default MITgcm-DARWIN formulation).

Where: PCᵐᵃˣⱼ = maximum carbon-specific growth rate for plankton j, γⁿᵘᵗⱼ = nutrient limition, γˡⁱᵍʰᵗⱼ = light limition, fᵗᵉᵐᵖⱼ = temperature limitation, γᶜᵒ²ⱼ = carbon dioxide limitation

\[\gamma^{light}_j = (1 - e^{(k_j^{sat} I)}) \cdot e^{k_{j}^{inh}} \cdot n_j^{light}\]

Agate.γⱼˡⁱᵍʰᵗ — Method

γⱼˡⁱᵍʰᵗ = (1 - ℯ^(kⱼˢᵃᵗ*I)) * ℯ^kⱼⁱⁿʰ * nⱼˡⁱᵍʰᵗ

Light limitation for plankton j (Default MITgcm-DARWIN formulation).

Where: kⱼˢᵃᵗ = half saturation constant of light saturation of plankton j, I = irradiance, kⱼⁱⁿʰ = half saturation constant of light inhibition of plankton j, nⱼˡⁱᵍʰᵗ = light penalty term of plankton j