Allometry

Agate.Library.Allometry.allometric_palatability_unimodal — Method

allometric_palatability_unimodal(prey_data, predator_data)

Calculates the unimodal allometric palatability of prey based on predator-prey diameters.

This function extracts prey_diameter, predator_diameter, optimum_predator_prey_ratio, and specificity from the provided dictionaries and calculates the palatability using the diameter-based formula.

Note that this formulation differs from the currently operational MITgcm-DARWIN model as it uses diameter instead of volumes and is structurally different. However, both formulations result in a unimodal response the width and optima are modulated by the optimumpredatorprey ratio and the specificity.

Arguments

prey_data: A dictionary containing prey-specific data:

diameters: Diameter of the prey.

predator_data: A dictionary containing predator-specific data:

can_eat: A binary value (1 or 0) indicating if the predator can consume prey. If this is set to 0, palatability is set to 0.

diameters: Diameter of the predator.

optimum_predator_prey_ratio: The optimal predator-prey diameter ratio for the predator.

specificity: A parameter controlling how sharply the palatability decreases away from the optimal ratio.

Returns

palatability: A number between 0 and 1 representing the palatability.

Agate.Library.Allometry.allometric_palatability_unimodal_protection — Method

allometric_palatability_unimodal_protection(prey_data, predator_data)

Calculates the unimodal allometric palatability of prey, accounting for additional prey protection mechanisms.

The function uses a modified unimodal relationship defined by: palatability = prey_protection / (1 + (predator_prey_ratio - predator_prey_optimum)^2)^predator_specificity

Arguments

prey_data: A dictionary containing prey-specific data:

diameters: Diameter of the prey.

protection: A scaling factor between 0 and 1 representing additional protection mechanisms of the prey.

predator_data: A dictionary containing predator-specific data:

can_eat: A binary value (1 or 0) indicating if the predator can consume prey. If this is set to 0, palatability is set to 0.

diameters: Diameter of the predator.

optimum_predator_prey_ratio: The optimal predator-prey diameter ratio for the predator.

specificity: A parameter controlling how sharply the palatability decreases away from the optimal ratio.

Returns

palatability: A number between 0 and prey_protection representing the palatability.

Agate.Library.Allometry.allometric_scaling_power — Method

allometric_scaling_power(a, b, d::Number)

Allometric scaling function using the power law for cell volume.

Arguments

a: scale

b: exponent

d: cell equivalent spherical diameter (ESD)

Growth

Agate.Library.Growth.default_PC — Method

PC = PCᵐᵃˣ * γⁿᵘᵗ *  γˡⁱᵍʰᵗ * fᵗᵉᵐᵖ *  γᶜᵒ²

Carbon-specific growth rate for plankton (Default MITgcm-DARWIN formulation).

Arguments

PCᵐᵃˣ: maximum carbon-specific growth rate

γⁿᵘᵗ: nutrient limition

γˡⁱᵍʰᵗ: light limition

fᵗᵉᵐᵖ: temperature limitation

γᶜᵒ²: carbon dioxide limitation

Mortality

Agate.Library.Mortality.linear_loss — Method

Linear mortality rate. In this formulation mortality is constant, and can be interpreted as a "closure term" for low density predation and and other death terms.

Arguments

P: plankton concentration

l: mortality rate

Agate.Library.Mortality.quadratic_loss — Method

Quadratic mortality coefficient. In this formulation mortality increases exponentially with plankton biomass and is often interpreted to represent viral processes and non-represented density-dependent predation effects.

Arguments

P: plankton concentration

l: mortality rate

Nutrients

Agate.Library.Nutrients.liebig_minimum — Method

liebig_minimum(nutrient_limitations)

Liebig's law of the minimum, which states that growth is limited by the scarcest (most limiting) resource.

Arguments

nutrient_limitations: an array of nutrient limitation values

Returns the minimum value among the given nutrient limitations.

Agate.Library.Nutrients.monod_limitation — Method

R / (kᵣ + R)

Monod formulation of nutrient limitation, which is based on Michaelis-Menten enzyme kinetics.

Arguments

R: nutrient (e.g. N, P, Si)

kᵣ: nutrient half saturation constant

Note that sometimes this formulation is also used for Predation ('Holling type 2').

Photosynthesis

Agate.Library.Photosynthesis.light_limitation_geider — Method

Pᶜₘₐₓ[1-exp((-αᶜʰˡθᶜE₀)/Pᶜₘₐₓ)]

A light limitation function which depends on the cellular ratio of chlorophyll to carbon. This formulation is based on equation (4) from Geider et al., 1998.

Arguments

PAR: photosynthetic active radiation (E₀)

maximum_growth_rate: maximum growth rate before nutrient limitation (Pᶜₘₐₓ)

photosynthetic_slope: initial photosynthetic slope (αᶜʰˡ)

chlorophyll_to_carbon_ratio: ratio between cellular chlorophyll and carbon (θᶜ)

Agate.Library.Photosynthesis.light_limitation_smith — Method

α * PAR / sqrt(μ₀ ^ 2 + α ^ 2 * PAR ^ 2)

Smith 1936 formulation of light limitation (also see Evans and Parslow, 1985).

Arguments

PAR: photosynthetic active radiation

α: initial photosynthetic slope

μ₀: maximum growth rate at T = 0 °C (this seems weird?, from Kuhn 2015)

Agate.Library.Photosynthesis.photosynthetic_growth_single_nutrient — Method

Single nutrient monod smith photosynthetic growth (used, for example, in Kuhn 2015).

Arguments

N: nutrient concentration

P: phytoplankton concentration

PAR: photosynthetic active radiation

μ₀: maximum growth rate at T = 0 °C

kₙ: nutrient half saturation

α: initial photosynthetic slope

Agate.Library.Photosynthesis.photosynthetic_growth_single_nutrient_geider_light — Method

Single nutrient geider photosynthetic growth.

Arguments

N: nutrient concentration

P: phytoplankton concentration

PAR: photosynthetic active radiation

maximum_growth_rate: maximum growth rate before nutrient limitation (Pᶜₘₐₓ)

kₙ: nutrient half saturation

photosynthetic_slope: initial photosynthetic slope (αᶜʰˡ)

chlorophyll_to_carbon_ratio: ratio between cellular chlorophyll and carbon (θᶜ)

Agate.Library.Photosynthesis.photosynthetic_growth_two_nutrients_geider_light — Method

photosynthetic_growth_two_nutrients_geider_light(
    DIN,
    PO4,
    P,
    PAR,
    maximum_growth_rate,
    half_saturation_DIN,
    half_saturation_PO4,
    photosynthetic_slope,
    chlorophyll_to_carbon_ratio,
)

Two nutrient geider photosynthetic growth.

Arguments

DIN: dissolved inorganic nitrogen concentration

PO4: phosphate concentration

P: phytoplankton concentration

PAR: photosynthetic active radiation

maximum_growth_rate: maximum growth rate before nutrient limitation

half_saturation_DIN: nitrogen half saturation

half_saturation_PO4: phosphate half saturation

photosynthetic_slope: initial photosynthetic slope

chlorophyll_to_carbon_ratio: ratio between cellular chlorophyll and carbon

Agate.Library.Photosynthesis.γˡⁱᵍʰᵗ — Method

γˡⁱᵍʰᵗ = (1 - ℯ^(kˢᵃᵗ*I)) * ℯ^kⁱⁿʰ * nˡⁱᵍʰᵗ

Light limitation for plankton (Default MITgcm-DARWIN formulation).

Arguments

I: irradiance

kˢᵃᵗ: half saturation constant of light saturation

kⁱⁿʰ: half saturation constant of light inhibition

nˡⁱᵍʰᵗ: light penalty term

Predation

Agate.Library.Predation.assimilation_efficiency_emergent_binary — Method

assimilation_efficiency_emergent_binary(prey_data, predator_data)

Determines the assimilation efficiency of a predator consuming prey, based on binary conditions of edibility.

The function evaluates whether the predator can eat the prey and whether the prey can be consumed, and assigns the assimilation efficiency accordingly.

Arguments

prey_data: A dictionary containing prey-specific data:

can_be_eaten: A binary value (1 or 0) indicating if the prey can be consumed by the predator.

predator_data: A dictionary containing predator-specific data:

can_eat: A binary value (1 or 0) indicating if the predator can consume prey.

assimilation_efficiency: The efficiency with which the predator assimilates nutrients from the prey if the conditions are met.

Returns

assimilation_efficiency:

If can_eat is 1 and can_be_eaten is 1, returns the predator's assimilation_efficiency.

Otherwise, returns 0.

Agate.Library.Predation.holling_type_2 — Method

Holling's "type II" functional response as describe in Holling 1959. The function is similar to the Monod equation and Michaelis-Menten equation of for enzyme kinetics. The formulation is characterized by decelerating predation as prey concentrations increase.

Arguments

R: prey density

k: prey density at which predation is half it's maximum rate

Agate.Library.Predation.predation_assimilation_loss_idealized — Method

Estimates the rate at which plankton predation gain is lost to the environment due to inefficient assimilation efficiency (e.g. 'sloppy feeding').

Note that this differs from the predationgainidealized as this function represents the transfer of biomass from the prey to the environment rather than the transfer of biomass from the prey to the predator.

Arguments

P: phytoplankton concentration

Z: zooplankton concentration

β: assimilation efficiency of prey to the predator

gₘₐₓ: maximum grazing rate of the predator

kₚ: grazing/holling half saturation

Agate.Library.Predation.predation_assimilation_loss_preferential — Method

Estimates the rate at which plankton predation gain is lost to the environment due to inefficient assimilation efficiency (e.g. 'sloppy feeding').

Note that this differs from the predationgainpreferential as this function represents the transfer of biomass from the prey to the environment rather than the transfer of biomass from the prey to the predator.

Arguments

P: phytoplankton concentration

Z: zooplankton concentration

β: assimilation efficiency of prey to the predator

gₘₐₓ: maximum grazing rate of the predator

kₚ: grazing/holling half saturation

palatability: the likelihood at which the predator feeds on the prey

Agate.Library.Predation.predation_gain_idealized — Method

Estimates the gain rate of Z (predator) feeding on P (prey). In this formulation predation rate is multiplied by an assimilation efficiency (β) which represents 'sloppy feeding'.

Arguments

P: phytoplankton concentration

Z: zooplankton concentration

β: assimilation efficiency

gₘₐₓ: maximum grazing rate

kₚ: grazing/holling half saturation

Agate.Library.Predation.predation_gain_preferential — Method

Estimates the gain rate of Z (predator) feeding on P (prey). In this formulation predation rate is multiplied by an assimilation efficiency (β) which represents 'sloppy feeding'.

Arguments

P: phytoplankton concentration

Z: zooplankton concentration

β: assimilation efficiency

gₘₐₓ: maximum grazing rate

kₚ: grazing/holling half saturation

palatability: the likelihood at which the predator feeds on the prey

Agate.Library.Predation.predation_loss_idealized — Method

Estimates the loss rate of P (prey), to Z (predator). In this formulation predator-prey interactions are modulated both by their density (Holling type 2) and the prey-predator palatability.

Arguments

P: phytoplankton concentration

Z: zooplankton concentration

gₘₐₓ: maximum grazing rate

kₚ: prey density at which predation is half it's maximum rate

Agate.Library.Predation.predation_loss_preferential — Method

Estimates the loss rate of P (prey), to Z (predator). In this formulation predator-prey interactions are modulated both by their density (Holling type 2) and the prey-predator palatability.

Arguments

P: phytoplankton concentration

Z: zooplankton concentration

gₘₐₓ: maximum grazing rate

kₚ: prey density at which predation is half it's maximum rate

palatability: the likelihood at which the predator feeds on the prey

Remineralization

Agate.Library.Remineralization.remineralization_idealized — Method

Idealized remineralization of detritus into dissolved nutrients.

Arguments

D: detritus

r: remineralization rate